minoe fix ve featured added

+++ e-commerce-backend/src/main/java/com/example/repositories/SubCategoryRelationRepository.java
@@ -11,7 +11,7 @@ import com.example.models.SubCategoryRelation;
 import java.util.List;
 
 @Repository
-public interface SubCategoryRelationRepository extends JpaRepository<SubCategoryRelation, Long> {
+public interface SubCategoryRelationRepository extends JpaRepository<SubCategoryRelation, Integer> {
     
     // Find all direct subcategories (depth=1) of a given category
     List<SubCategoryRelation> findByAncestorAndDepth(Category ancestor, int depth);
